Remove cancel multiples from new API
